hci: use 32-bit bit to store local supported commands

+++ src/hci.h
@@ -958,7 +958,7 @@ typedef struct {
     uint8_t local_supported_features[8];
 
     /* local supported commands summary - complete info is 64 bytes */
-    uint16_t local_supported_commands;
+    uint32_t local_supported_commands;
 
     /* bluetooth device information from hci read local version information */
     // uint16_t hci_version;
